Score System (v1.1.1)
Added a basic score system with saving capabilities to the game. Other save data could be stored in the BP_Save_Data.uasset but this is for further implementation. Does have to be coupled with the final character after player movement is finished and with the chunk generator when that is finished.